Cast scalar fields to scalar and related refactoring.
Changes:

The function _sanitize_data() now converts to scalar
all the arrays or sequences with lenghth == 1 that have
"scalar" type in the specs. This assure consistency
in the created files. As before, it casts the type
of detectors_specs to match detectors array.

save_photon_hdf5() now calls assert_valid_photon_hdf5()
(with argument sanitize=True), so that the input dict
is first sanitized (modified in-place) and then checked
against the specs.

These changes increase consistency and robustness of
save_photon_hdf5() regardless of the variability in
the input data.