Ensure h5file is closed if requested
Callers of the save_photon_hdf5() can specify that they wanted the hdf5
file they passed in closed before returning. However, we performed
validation before closing the file, and the validation can error out. If
this happened, the file /would not be closed/. Not a bit deal if the
user still had a hold on the file object and could close it themselves,
but if they instead passed in just the file name then they were unable
to close it themselves without somehow getting a hold of the object from
python's global list of open files. This meant that they would need to
restart the python process before they could try saving to the hdf5 file
again.

Now, we place the file closing logic in the clause of a finally:
statement to ensure we always close it if the user wants us to.